Create a specefic css for each search process
